💄 `/sutras/:id`をレスポンシブデザインに対応する
