+++ frontend/src/pages/sutras/[id].tsx
@@ -1,18 +1,14 @@
 import axios from 'axios'
 import { GetServerSideProps, GetServerSidePropsContext } from 'next'
 import { getSession } from 'next-auth/react'
-import { useState } from 'react'
 import CustomHead from '@/components/CustomHead'
 import Footer from '@/components/Footer'
 import Header from '@/components/Header'
 import { railsApiUrl } from '@/config/index'
-import Map from '@/features/map/Map'
-import Note from '@/features/note/Note'
-import CapturedImage from '@/features/photo/CapturedImage'
-import DeletePhoto from '@/features/photo/DeletePhoto'
-import EditPhoto from '@/features/photo/EditPhoto'
 import PhotoUploadAndPreview from '@/features/photo/PhotoUploadAndPreview'
 import fetchPhotoId from '@/features/photo/fetchPhotoId'
+import SutraDetailsMdLayout from '@/features/sutra/SutraDetailsMdLayout'
+import SutraDetailsOverMdLayout from '@/features/sutra/SutraDetailsOverMdLayout'
 import fetchUserId from '@/features/user/fetchUserId'
 import { SutraDetailsProps } from '@/types/types'
 
@@ -81,85 +77,6 @@ export const getServerSideProps: GetServerSideProps = async (
 }
 
 const SutraDetails = ({ sutra, photo }: SutraDetailsProps) => {
-  const currentLocation = {
-    lat: photo.latitude,
-    lng: photo.longitude,
-    img: photo.cropped_image_url,
-  }
-  const [editNote, setEditNote] = useState(!photo.note)
-  const [editMode, setEditMode] = useState(false)
-
-  const renderNote = () => {
-    if (editNote) {
-      return (
-        <Note
-          photoId={photo.id}
-          sutraId={sutra.id}
-          photoNote={photo.note}
-          setEditNote={setEditNote}
-        />
-      )
-    } else {
-      return (
-        <div className='flex flex-col justify-center items-center '>
-          <div className='w-80 h-32 rounded-lg bg-gray-300 overflow-hidden'>
-            <div className='mx-2 my-1'>{photo.note}</div>
-          </div>
-          <div className=' bg-blue-500 hover:bg-blue-400 text-white rounded-full font-notoSans text-xs  ml-auto my-2 px-2 py-1'>
-            <button onClick={() => setEditNote(true)}>メモ編集</button>
-          </div>
-        </div>
-      )
-    }
-  }
-
-  const renderSutraDetails = () => {
-    if (editMode) {
-      return (
-        <div>
-          <PhotoUploadAndPreview sutraId={sutra.id} photoId={photo.id} sutra={sutra} />
-        </div>
-      )
-    } else {
-      return (
-        <div className='flex flex-col justify-center items-center'>
-          <div>
-            <CapturedImage
-              capturedImageUrl={photo.image_url}
-              width={360}
-              height={360}
-              borderRadius='5px'
-            />
-          </div>
-          <div className='flex justify-center mt-8'>
-            <div className='text-5xl text-black/25 font-kinuta mx-4'>{sutra.kanji}</div>
-            <div className='mx-4 mt-1'>
-              <CapturedImage
-                capturedImageUrl={photo.cropped_image_url}
-                width={48}
-                height={48}
-                borderRadius='5px'
-              />
-            </div>
-          </div>
-          <div className='mt-8 mb-2'>
-            <Map markerLocation={currentLocation} />
-          </div>
-          <div className='mb-8'>住所：{photo.address}</div>
-          <div>{renderNote()}</div>
-          <div className='flex justify-center content-between'>
-            <div className='bg-blue-300 hover:bg-blue-200 text-gray-700 rounded-full font-notoSans text-sm mx-4 my-8 px-4 py-2'>
-              <EditPhoto setEditMode={setEditMode} />
-            </div>
-            <div className=' bg-gray-400 hover:bg-gray-300 text-white rounded-full font-notoSans text-sm mx-4 my-8 px-4 py-2'>
-              <DeletePhoto photoId={photo.id} />
-            </div>
-          </div>
-        </div>
-      )
-    }
-  }
-
   return (
     <div className='flex flex-col min-h-screen bg-beige'>
       <CustomHead title='詳細' />
@@ -170,7 +87,14 @@ const SutraDetails = ({ sutra, photo }: SutraDetailsProps) => {
             <PhotoUploadAndPreview sutraId={sutra.id} photoId={photo.id} sutra={sutra} />
           </div>
         ) : (
-          renderSutraDetails()
+          <div>
+            <div className='md:hidden'>
+              <SutraDetailsMdLayout sutra={sutra} photo={photo} />
+            </div>
+            <div className='hidden md:flex'>
+              <SutraDetailsOverMdLayout sutra={sutra} photo={photo} />
+            </div>
+          </div>
         )}
       </main>
       <Footer />
